fix extra map3

+++ SuperMarioBros3/Animations.cpp
@@ -26,7 +26,7 @@ void CAnimation::Add(int spriteId, DWORD time)
 // NOTE: sometimes Animation object is NULL ??? HOW ??? 
 void CAnimation::Render(float x, float y, int alpha)
 {
-	DWORD now = GetTickCount64();
+	ULONGLONG now = GetTickCount64();
 	if (currentFrame == -1)
 	{
 		currentFrame = 0;
