+++ SuperMarioBros3/main.cpp
@@ -114,7 +114,7 @@ int Run()
 {
 	MSG msg;
 	int done = 0;
-	DWORD frameStart = GetTickCount64();
+	ULONGLONG frameStart = GetTickCount64();
 	DWORD tickPerFrame = 1000 / MAX_FRAME_RATE;
 
 	while (!done)
@@ -127,8 +127,8 @@ int Run()
 			DispatchMessage(&msg);
 		}
 
-		DWORD now = GetTickCount64();
-		DWORD dt = now - frameStart;
+		ULONGLONG now = GetTickCount64();
+		ULONGLONG dt = now - frameStart;
 
 		if (dt >= tickPerFrame)
 		{
